rename: fisheye_warping_g1.py -> fisheye_warping_tiled.py
Better reflects what the module actually is (a tiled-cubemap renderer)
rather than tying the name to G1 specifically.
